Read.me
Android Fundamentals in Kotlin Solution Apps
===========================================
This repository contains the solution code for the apps for the
Android Fundamentals in Kotlin course, published by the Google Developer Training Team.
See [Android Training](https://developer.android.com/courses).
A “solution app” is an app that contains the code for a completed codelab.
Code for the starter apps in this course are in a different repository.
Introduction
------------
Each of the samples in this repository is associated with a codelab exercise
from the Android Fundamentals in Kotlin course.
The Android Fundamentals in Kotlin course can be used for self-study, or as the 
basis of an instructor-led training. It is targeted towards beginning Android 
developers with a background in software development or computer science with 
at least two years of experience with a modern, object-oriented programming
language, such as the Java programming language,
and familiarity with basic software architecture principles. 
Since this course is in Kotlin, it is assumed you know Kotlin at the level 
taught in the Udacity course, 
[Kotlin Bootcamp for Programmers](https://www.udacity.com/course/kotlin-bootcamp-for-programmers--ud9011).
Pre-requisites
--------------
The samples, along with the course, assume basic competence in software
design and development, as well as some background in computer science.
Specifically, to get started you need:
- Familiarity with the general software development process for object-oriented
applications using an IDE (Integrated Development Environment).
- At least 1-3 years of experience with object-oriented programming and
the Java programming language, or a comparable modern language. 
- Familiarity with the Kotlin programming language.
Android Studio Version
----------------------
During development of this course, all samples were built using Android Studio 3.3,
and the codelabs assume you are using Android Studio 3.3 with corresponding libraries
and available features. See the 
[Android Studio download page](https://developer.android.com/studio) for details.  
Getting Started
---------------
1. [Install Android Studio](https://developer.android.com/studio/install.html),
if you don't already have it. (If you are just starting the course, you will be guided
through this process.)
2. Download the starter sample for the codelab.
2. Open the sample into Android Studio.
3. Build and run the sample. You may need to update gradle and library versions. 
Follow the guidance provided by Android Studio.